Raise context of Document component.
+ Move Document component definition into Dir component.
+ Move blog and notes posts into _posts directory.
+ Move blog and notes splash page into blog and notes directory.
