+++ _posts/blog/Test_Blog.md
@@ -1,5 +1,5 @@
 ---
 title: 'Hello, World!'
-date: 'Sep 11, 2022'
+date: '2022-09-01'
 ---
 Hello, World!
