--- styles/Document.module.css
@@ -1,61 +1,0 @@
-.document {
-  padding: 5px;
-  display: flex;
-  flex-direction: row;
-  align-content: flex-start;
-  gap: 10px;
-
-  min-height: 70px;
-  width: 365px;
-
-  background-color: rgba(40, 42, 54, 0.3);
-  border-radius: 10px;
-}
-
-.document:hover {
-  box-shadow: 2px 2px 5px #282a36, -2px 2px 5px #282a36, 2px -2px 5px #282a36, -2px -2px 5px #282a36;
-}
-
-.file {
-  font-size: 2.6em;
-  margin: 0 auto;
-}
-
-.details {
-  flex: 1;
-
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-
-  height: 100%;
-  width: 80%;
-  font-size: 0.75em;
-
-  overflow-wrap: break-word;
-}
-
-.description {
-  width: 100%;
-}
-
-.date {
-  color: #6272a4;
-  font-style: italic;
-}
-
-.tags {
-  width: 100%;
-  word-break: normal;
-}
-
-.tag {
-  border: 1px solid black;
-  border-radius: 24px;
-  padding: 2px;
-}
-
-.postlink {
-  color: inherit;
-  text-decoration: inherit;
-}
